requirements.txt
a2a-sdk==0.3.24
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.131.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==2.19.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.2.0
ipython==8.38.0
jedi==0.19.2
jiter==0.13.0
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==26.0
parso==0.8.6
platformdirs==4.9.2
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
vertexai==1.71.1
wcwidth==0.6.0
